LPD-48223 Extract MarketplaceViews to reuse it
